fix(v4): move k2 dispatch tests to sibling file (≤500 LoC cap)
Relocate the k2_dispatch tests into forward_k2_dispatch_tests.rs via #[path]
so forward_k2.rs stays within the ≤500-LoC-per-file CI cap. No logic change.
